Migration for VmHost allow_slices field
We are introducing a new column to vm_host model to select hosts that will be used to place VMs running inside slices. We want to keep the slice or no-slice logic separate per-host, to avoid confusion and difficulty in allocating and tracking cpus and memory. Therefore each host will either continue allocating VMs running unconstrained, as it was so far, or will allocate VMs running inside slices. The value of the accepts_slices column will decide that.