Migration for cpu limits on a vm
Adding two columns that track the cpu percent limit that a VM is allowed to use from the overall total of cpu percentage on a host. At the host the total is equal to #cpus * 100%. So for a 96 cpu host, that is 9600% of cpu. Each VM is given a portion of that. For example standard-2 uses 200% out of 9600% of total cpu on a host.
The second column tracks the amount of cpu percent that a VM is allowed to burst into. This will be used for burstable VMs introduced in subsequent commits.

Combining those two for an ease of deployment.