Read Load Balancer Node state according to threshold
We are updating load balancer node when the defined thresholds are
passed, but a flipping health check, or an unrelated update may simply
declare nodes unhealthy when they are just in the verification phase
(state_count < threshold).

This commit adds the new node_state function to better understand the
status. We basically smoothen the state read curve.