Add ipv4 routes for cross vm packets
So far, VMs had a /32 ip range and VMs did not need to route
packets for other entities with ip inside. For Kubernetes CNI,
we will have pods which need to send packets out of the VM to another
destination in the same private subnet. In that case we need to add routes
to the VM's network namespace.