Decrease neg-ttl in dnsmasq config
Default neg-ttl is 300 seconds which is a lot to wait for retrying
dns queries. For bootstrapping kubernetes, we have 4 minute timeout
to discover the apiserver through the loadbalancer and DNS. If the
first query fails, the bootstrap will fail and way more complex
measurements should be added to resolve it.

In this commit the neg-ttl is set to 30 seconds which is reasonable for
retrying failed dns queries