Add support for vm ssh CLI command
This supports automatically connecting to the requested VM via SSH.
It takes options for which address family to use, and to override the
username used when connecting.

This passes information on the command the client should execute
via the headers understood by bin/ubi.  This includes support for
letting the client know which part of ARGV should be used in the
SSH connection, supporting commands such as:

  vm ssh location vm-name remote-command remote-command-arg

Note that users must provide both the location and vm name or
ubid when connecting.  This is because a principle of UbiCli is
that it does not access the database or use the models, it only
makes internal requests to Clover's api, and the api URLs require
the location and either the name or ubid.