Adding diagnostics to Allocator
Adding a way to dump the whole Allocator main query into a sql file, with all parameters inline, such that it can be run directly against a server. This can help troubleshooting the allocation problems, specifically "no capacity" errors, and also can aid in development. The logic is guarded by a Project feature flag, so it can be enabled in production for specific projects only.