Add State for Kubernetes Cluster
The state is displayed in the list and details views.
We have 3 states: creating, running and deleting.

The granular states for individual VMs displayed in the details view are
not super correct, as a node will be in "running" state even though the
k8s is being configured or failed within the VM. That's a patch for
another day.