Enforce restriction that UbiCli/UbiRodish do not query the database
This uses the new block_queries extension in Sequel, to block
queries around the call to UbiCli.process, and only allow them
when UbiCli internally calls Clover.call.

With this, if UbiCli/UbiRodish tries to access the database
directly, it will fail.

This restriction is only added when running frozen specs, so that
code must pass both with and without the restriction (useful as
the restriction is not used in production).