Update Load Balancer for newly added VMs even when they are down
We were waiting for the first "up" signal to update the load balancer
node. However, the moment we add a new node, we update the dns records
and this node starts getting traffic. If we do not update the load
balancer nodes, this node simply doesn't start rerouting the traffic to
already existing healthy nodes. Therefore, we must immediately update
the load balancer nodes.

We were also not cleaning up the node NAT settings for a detached node
which continues to reroute the traffic silently. We start removing it.