Add more data to the deadline pages
Currently, the deadline pages only show the resource type, ID, and
expected target label. It would be helpful to include additional data
based on the resource type for the on-call engineer. For example, when a
host crashes, we receive multiple deadline pages from virtual machines,
making it easier for the on-call engineer to identify the host by
reviewing the incident details in Slack.

Additionally, while I haven't implemented this, we could choose whether
or not to trigger the page based on the resource type to avoid page
storms. For instance, when creating a new deadline page for a virtual
machine, we can check if its host is unavailable. If the host is down,
we can decide not to trigger the page for the virtual machine.