Prepare load_balancers_vms models for health probes
We are moving the health checks into load_balancers_vms. The main reason
these are not handled as part of VM monitoring is because the VM might
be up and running but the load balancer makes a decision based on the
application state. It is safer to keep this separate. This code will
also interract with load balancer and keeping a code path as part of a
VM that is not part of a load balancer doesn't seem clever. We are
making a couple of logic changes here.
1. The state of the load balancer node is only updated when the monitor
is sure their state is changed. This way, we don't have to look into
counters at the time of update.
2. IPv4 and IPv6 availability is handled properly. Before, these were
kind of mixed up and causing some logical errors. Now, they are
distinctly checked and properly declared.

This commit itself doesn't simply stop the health probe strands. We will
need to handle their migration manually by destroying the strands.