Add the target's ubid to SSH command logs
Our logs include the ubid of the strand that initiated the command as
the `thread` property. However, this ubid refers to the initiator, not
the target.

For example, when a GitHub runner strand runs SSH commands on both the
VM host and the VM, the `thread` for both is the same. If the host
experiences SSH latencies, we can't aggregate the channel durations of
the VM host by `thread`. Adding the sshable's `ubid` will help us
aggregate the durations for the target. This way, if a specific host has
SSH latencies, we can monitor it effectively.