Simple stress script
Adds a script to run `stress-ng` workload on one or more VMs.
Usage:
- `stress_up --install` - installs `stress-ng` on all VMs
- `stress_up` - runs the stress-ng workload on all VMs
- `stress_up <vm_name1> <vm_name2>` - runs the stress workload on specified VMs. This also works when combined with the --install option.

The command to run is specified inside the script as `STRESSNG_LOAD_COMMAND` and can be easily tweaked.
After the run the script summarizes the load on each VM in a simple table and also computes the avarage from all VMs.