Change allocator diagnostics to use structured output
Normally, we do not encode dynamic information in the "banner string"
given to `Clog.emit`, it is static.  Visit
f2b128f972ce81847956e386d626bd9a94c0106e for the underpinnings of
Clog.

It was also unnecessary to introduce a coverage gap here: the result
is, once diagnostics is enabled, the process could run unchecked code
and crash.  The structured output also makes it easy to check some
invariants about the printed data.

There is one regression, however: the old way of doing things would
look better in plain text, for copy-and-paste.  But the new way would
only look good in something that de-escapes JSON, in particular
double-quotes for SQL identifiers, and renders it, such as the log
analyzer in the production system.  I am not sure how to remediate
this.