Suppress "void context" warnings
When running specs, I get output like this at first, and quite a few
times with `rake coverage_pspec`:

    clover/loader.rb:177: warning: possibly useless use of :: in void context

Previously, there were pragmas here to squelch Rubocop's linting of
this, and that was adequate, these warnings are from Ruby, so it
doesn't work for that.

Ironically, Enes first proposed a version like this, and I suggested
that a version that was statically simpler to analyze (for people)
would be preferred, to use the void-context expressions with a
suppressing Rubocop pragma, but I guess our hand is forced now anyway.