CertNexus restart counter is added to not DOS ZeroSSL
In case the passed DNSZone is not valid globally, DNS validation or cert
generation may fail. In that case, CertNexus restarts to do a new acme
dance. It goes on forever. Instead of this, we are introducing a restart
logic. In case of restarts, we start napping aggressively up until 10
minutes.