Fixing race condition in E2E tests
In both `Test::ConnectedSubnets` and `Test::FirewallRules` the test is installing a listening service on one of the VMs and connects to it from other VMs. However the method that returns that first VM could return different instances each time it runs, as it relies on the DB returning the VM instances in the same order each time it is called.

I added a way to preserve the ID of the VM where the listener service is installed, and use that to retrieve the same instance later.