Adding E2E tests for slices
Updating E2E tests to validate VM creation inside a slice. The test switches the host's `accept_slices` flag to true and sets the project flag to use slices, then runs the standard VM test to create VMs and test them. Those VMs are creted inside slices on a host. An additional verification is added into VmGroup to check the validity of the slices - `VmHostSlices`.