Attempt to recreate VM up to 2 times if it fails during prep
Previously, if a VM failed during Vm::Nexus#prep, it would continually
be stuck in that step, because #prep treated Failed the same as NotStarted,
and the steps to initially start prep are not the same as the steps to
recover from failure.  I'm not sure what the steps are to recover from
failure (since there are many different things that could fail), so this
takes the hammer approach.  It sets a recreate semaphore and hops to
destroy.

When destroy (#destroy_slice) is called, if the recreate semaphore is set,
after destroying the VM, it assembles a new VM with the same ubid, and
then hops to start instead of exiting.

To prevent infinite recreation, there is a hard coded limit of 3 attempts.
After 3 failed attempts, the Strand is moved to prep_failed state, where
it stays until manual intervention. This generates a page to ensure it
isn't missed.

To implement this, I added the following keyword arguments to
Vm::Nexus#assemble:

* ubid: ubid to set on the created Vm
* vm_size: a VmSize object, which can be used directly instead of
  determining it from the size and arch arguments
* attempt: The number of attempts made to create the VM, defaults to 1
* strand: Updates existing strand instead of creating a new strand

To hold more state, it adds some additional data to the strand's stack,
so that the destroy_slice method can pass it back to assemble.

Other minor changes:

* Add Vm#vm_size. Extracted from Vm#display_size, this is used to set
  the vm_size argument to assemble.

* Do not show "deleting" as the display state for a Vm with recreate
  set, as showing that would look weird to the user.

* If requesting to create a Vm with a specific Nic (passing nic_id
  to assemble), when destroying and recreate is set, the related
  Nic is not destroyed when the Vm is destroyed, as it will be
  used for the newly created Vm.  I'm not sure this is used
  outside the tests, so it may be overkill.

* Remove transaction around Vm::Nexus#final_clean_up. This isn't
  needed, as it is only called inside a label, and labels have
  an implicit transaction around them.