Add migration for kubernetes vm size specification
This patch adds new columns to kubernetes models so that we allow
specific compute and storage sizes on Kubernetes VMs.

This change applies to both KubernetesCluster model for control plane
nodes and KubernetesNodepool model for worker nodes.

For both models, the compute specification is not null and the default
will be set in the nexus.

The theory is that, kubernetes community usually are
not interested in the actual size of the nodes, and slap cluster
autoscaler to handle the number of nodes.

Even less important, is the storage size of the nodes, so
that's left as null-able. In clover, we usually pick the lowest available
storage size of the selected compute size by default, so we can implement
the same logic for k8s nodes as well.