Don't wait to collect telemetry for the runner when for all SSH errors
We save runner script logs and serial logs if the job isn't successful
for further investigation. We also collect DockerHub limits to monitor
usage.

If the host or VM is unavailable, it raises an exception other than
Sshable::SshError, causing the runner program to get stuck.

If they are down, we shouldn't wait for the runner to collect telemetry.

I also extracted it to a helper method, which made the tests more
readable.