Fix destroy labels for the vm and inference endpoint
If the resource is already being destroyed, we don't want to hop to the
`destroy` label again when the `destroy` semaphore is incremented.

We check the current label to avoid hopping again. Some resources have
labels after the `destroy` label, so we need to check those as well.

We checked this for some resources, but it seems we missed it for the vm
and inference endpoint.