Support coverage for specific rhizome files
When rspec is run against rhizome folder with COVERAGE=rhizome,
The line coverage and enforcement can be observed, similar to
our CP test suite.

Currently, only the rhizome/kubernetes/lib/ubi_cni.rb is included
in the tracked files for rhizome, other files need to be added manually.

The change is included in ci.yml too, so that once a file makes into
the tracked rhizome files array, it will always be enforced.