Do not trigger a deadline page if the host is unavailable
When the host crashes, all resources on it trigger a deadline page,
which creates a page storm. This is overwhelming for the on-call
engineer and serves no purpose. If the VM host is unavailable, there’s
no need to trigger a deadline page.

Some resources don't have separate unavailability page type and instead
use deadlines to create them. Because of this, even if the VM host is
unavailable, we create a deadlines page if the current label is
"unavailable".

Our VM host monitoring isn't very sufficient right now. Adding disk
monitoring and possibly SPDK service monitoring will make this PR more
useful.