Clover: Support bdev_ubi 0.3 & increase SPDK hugepages.
bdev_ubi 0.3 fixes several error handling issues, mainly the ones
related to when we are out of iobuf free items which caused crashes in
production.

When we added support for bdev_ubi 0.3 in the Clover side, we increased
number of hugepages used for SPDK. So, this PR also increases that on
the Clover side.