Add cli vm sftp and vm scp commands
The vm sftp command is fairly similar to the vm ssh command, the main
difference is it takes a fixed number of arguments:

  ubi vm sftp location vm-name

The vm scp command is more complex, because transfers can be either
remote-to-local or local-to-remote.  I choose to not allow
local-to-local or remote-to-remote scp styles due to complexity.
Because you do not know the host in advance, I decided to have an
initial colon in the argument specify remoteness.  So the commands
look like:

  ubi vm scp location vm-name local-file :remote-file
  ubi vm scp location vm-name :remote-file local-file

The bin/ubi code could not handle the local-to-remote case, because
it cannot use ubi-command-argv-tail, as the argument it needs to
provide must come second.  I decided to add ubi-command-argv-initial
to specify a single element of argv to include as a command line
argument, before the argument specified by ubi-command-arg.

While here, I decided to use Process.exec when execing the command,
which avoids the need for the explicit exit.  For consistency with
the error paths, I also had the success path set an explicit exit.

As the ssh, sftp, and scp code share a lot of code, I refactored
the shared code into a UbiCli#handle_ssh method and
UbiCli::SSHISH_OPTS constant.