Remove the version selection from k8s creation form
Since we are a new service, let's not deal with allowing an older
version of k8s on the form. If someone asks for it, we can create
v1.31 on the backend anyway. Otherwise, everything is v1.32.

This is a leftover item that fell through the cracks during the
lengthy review process.

We expect v1.33 to be released in a short while and will bring this
back again.