Remove Subnet Selection From k8s Creation Form
With this patch, we remove the Subnet Selection from Kubernetes
creation form. Instead, we create a default subnet with a larger
ipv4 prefix (/16 instead of /26) with the name "cluster-k8s-subnet".

This way, the flow is simplified, by not requiring customer to create
a subnet before creating a cluster. Also, the subnet would need to have
a larger prefix and changing that is not easy at the moment.