Fix private subnet selection in a way to support larger subnets (/16)
Previously, random_private_ipv4 function could randomly error out
when requesting larger subnets, because one of the classes were not
large enough to accomodate /16 and the code didn't handle this case.

Now we retry the function if the initially selected family of large subnet
cannot accomodate the requested cidr size.

This is needed when we request bigger subnets for kubernetes clusters.