Switch cli vm ssh/sftp/scp from pre to post subcommands
New format of commands:

  vm location vm-name ssh
  vm location vm-name sftp
  vm location vm-name scp local-path :remote-path

  # ssh command with arguments
  vm location vm-name ssh cmd cmd-arg

  # With options
  vm location vm-name ssh -- -A --
  vm location vm-name ssh -- -A -- cmd cmd-arg
  vm location vm-name sftp -- -A
  vm location vm-name scp local-path :remote-path -r

Note the use of -- for ssh/sftp, to separate the options parsed by
UbiCli, from the options that will be passed directly to ssh/sftp
on the command line.  ssh actually needs 2 sets of --, because it
also uses -- to separate the ssh command line options from the
ssh command line arguments.